Extract and deduplicate chat effect interfaces
